requirements/base.txt
Markdown>=2.4.1
django-model_name-utils>=2.2
bcrypt>=1.0.2
psycopg2>=2.5.3
Pillow>=2.8.1
